Move scan to jetpack package
