+++ WordPress/src/main/AndroidManifest.xml
@@ -643,7 +643,7 @@
 
         <!--Scan-->
         <activity
-            android:name=".ui.scan.ScanActivity"
+            android:name=".ui.jetpack.scan.ScanActivity"
             android:label="@string/scan"
             android:theme="@style/WordPress.NoActionBar"/>
